refactor: update module path and dependencies
- Updated module path from gocasts/ToDoApp to github.com/Shayan-Ghani/GOAD
- Updated go.mod and go.sum with latest dependency versions
- Migrated import paths across all files to use new module path
- Added config package to replace previous constant package
- Simplified configuration management
