+++ config/config.go
@@ -1,3 +1,3 @@
-package constant
+package config
 
 const DefaultDSN = "root:root@tcp(localhost:3306)/todo_dev?multiStatements=true"
