+++ go.mod
@@ -1,15 +1,15 @@
-module gocasts/ToDoApp
+module github.com/Shayan-Ghani/GOAD
 
 go 1.22.1
 
 require (
 	github.com/go-sql-driver/mysql v1.8.1
-	github.com/golang-migrate/migrate/v4 v4.18.1
+	github.com/golang-migrate/migrate/v4 v4.18.2
 )
 
 require (
 	filippo.io/edwards25519 v1.1.0 // indirect
 	github.com/hashicorp/errwrap v1.1.0 // indirect
 	github.com/hashicorp/go-multierror v1.1.1 // indirect
-	go.uber.org/atomic v1.11.0 // indirect
+	go.uber.org/atomic v1.7.0 // indirect
 )
