+++ internal/delivery/cli/command/item.go
@@ -3,10 +3,10 @@ package command
 import (
 	"flag"
 	"fmt"
-	"gocasts/ToDoApp/internal/repository"
-	"gocasts/ToDoApp/pkg/formatter"
-	"gocasts/ToDoApp/pkg/response"
-	"gocasts/ToDoApp/pkg/validation"
+	"github.com/Shayan-Ghani/GOAD/internal/repository"
+	"github.com/Shayan-Ghani/GOAD/pkg/formatter"
+	"github.com/Shayan-Ghani/GOAD/pkg/response"
+	"github.com/Shayan-Ghani/GOAD/pkg/validation"
 	"time"
 )
